Extract shared FolderNavigableList and deprecate flat view (#2557)
Extract common folder-navigation logic (tree building, breadcrumbs,
Paginate wiring) into a reusable FolderNavigableList component. Both
FileBrowser and ExistingFileSelector now use it as thin wrappers,
removing ~370 lines of duplication. Flat view is removed in favour of
folder view only. ExistingFileSelector becomes a controlled component
with files passed as props. Release file browser is now always
read-only and the unused mode prop is removed from
ReleaseAssetsAccordion.
