+++ frontend/src/entry/model/releases/ReleaseDisplay.tsx
@@ -68,12 +68,7 @@ export default function ReleaseDisplay({
             <ReleaseAssetsMainText model={model} release={release} latestRelease={latestRelease} />
             <Box>{(release.files.length > 0 || release.images.length > 0) && <Divider />}</Box>
             <Stack spacing={1}>
-              <ReleaseAssetsAccordion
-                model={model}
-                release={release}
-                mode='interactive'
-                hideFileDownloads={hideFileDownloads}
-              />
+              <ReleaseAssetsAccordion model={model} release={release} hideFileDownloads={hideFileDownloads} />
               <ReleaseAssetsResponses model={model} release={release} />
             </Stack>
           </Stack>
